+++ crypto-onramp/src/main/java/com/stripe/android/crypto/onramp/exception/OnrampErrorMapper.kt
@@ -3,7 +3,6 @@ package com.stripe.android.crypto.onramp.exception
 import android.content.Context
 import com.stripe.android.core.StripeError
 import com.stripe.android.core.exception.StripeException
-import com.stripe.android.core.version.StripeSdkVersion
 import com.stripe.android.crypto.onramp.R
 import com.stripe.android.crypto.onramp.analytics.OnrampAnalyticsEvent
 
@@ -11,13 +10,14 @@ internal fun Throwable.toCryptoOnrampError(
     context: Context,
     operation: OnrampAnalyticsEvent.ErrorOccurred.Operation,
     publishableKey: String?,
+    additionalSdkVersions: List<SDKVersion> = emptyList(),
 ): Throwable {
     if (this is StripeCryptoOnrampError) return this
 
     val stripeException = this as? StripeException ?: return this
     val stripeError = stripeException.stripeError ?: return this
     val apiUserMessage = stripeError.extraFields?.get(FIELD_USER_MESSAGE)?.takeIf { it.isNotBlank() }
-    val sdkVersion = StripeSdkVersion.VERSION
+    val sdkVersions = listOf(SDKVersion.stripeAndroid) + additionalSdkVersions
 
     val apiErrorContext = APIErrorContext(
         reason = stripeError.extraFields?.get(FIELD_REASON),
@@ -35,7 +35,7 @@ internal fun Throwable.toCryptoOnrampError(
     return if (stripeError.isAppAttestationError()) {
         AppAttestationException(
             context = apiErrorContext,
-            sdkVersion = sdkVersion,
+            sdkVersions = sdkVersions,
             fallbackUserMessage = context.getString(
                 R.string.stripe_onramp_app_attestation_default_user_message,
             ),
@@ -43,7 +43,7 @@ internal fun Throwable.toCryptoOnrampError(
     } else {
         UncategorizedApiErrorException(
             context = apiErrorContext,
-            sdkVersion = sdkVersion,
+            sdkVersions = sdkVersions,
             fallbackUserMessage = context.getString(
                 R.string.stripe_onramp_default_api_error_user_message,
             ),